backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
google-auth==2.37.0
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1
